fix(web/applications): use file extension if type is not determined
The type of a file is determined by the browser and accessed using the
`type` property of the file object. For more obscure file types this can
fail and you end up with an empty string. This change allows the
extension to be used if the real type isn't determined correctly.