refactor(packages/redis): use classes for MSAL plugin configuration
This is the way the documentation suggests to initialise an MSAL plugin.
Also, the MSAL library expects the Redis client to always return a
string, never null as is default in the redis npm package. The custom
MSALCacheClient class gets around this by returning an empty string if
the true result is null.